split up windows functions into a dedicated module with separate files (#724)
* split up window functions

* add docs
